+++ lib/karafka/setup/configurators/water_drop.rb
@@ -6,6 +6,7 @@ module Karafka
       # Class responsible for setting up WaterDrop configuration
       class WaterDrop < Base
         # Sets up a WaterDrop settings
+        # @param config [Karafka::Setup::Config] Config we can user to setup things
         def self.setup(config)
           ::WaterDrop.setup do |water_config|
             water_config.deliver = true